pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "cah"
version = '2.2.0'
description = "A slackbot for playing Cards Against Humanity on Slack."
authors = ["bobrock <bobrock@tuta.io>"]
license = "MIT"
readme = 'README.md'
repository = 'https://github.com/barretobrock/cah_bot'
packages = [
    { include = 'cah' },
]
include = ["CHANGELOG.md"]

[tool.poetry.dependencies]
python = "^3.10"
slacktools = {git = "https://github.com/barretobrock/slacktools.git"}
pukr = { git = "https://github.com/barretobrock/pukr.git", branch = "main" }
Flask = "^3"
Flask-SQLAlchemy = "^3"
numpy = "^1"
pandas = "^2"
psycopg2 = "^2"
pykeepass = "^4"
requests = "^2"
slack_bolt = "^1"
# Optional dependencies would go down here
# example = { version = ">=1.7.0", optional = true }

[tool.poetry.dev-dependencies]
pre-commit = "^3"
pytest = "^7"
pytest-cov = "^4"
flake8 = "^6"
tox = "^4"

[tool.isort]
profile = 'black'
force_grid_wrap = 2
force_sort_within_sections = true
skip_gitignore = true
skip_glob = ['*_pb2.py', '*_pb2_grpc.py']

[tool.poetry.extras]
test = ["pytest"]
